Add files via upload
We used wavelength calibration to correspond the pixels of our detectors to the wavelength
that they detect and calculate the temperature of our blackbody source. We then
came up with a wavelength solution for a 2-dimensional dispersed image using OH skylines
to estimate the velocity of an ionized iron gas of a supernova explosion. Finally, we extracted
the spectrum of 3 stars and analyzed their spectral graphs to predict their temperature and
motion